+++ src/Driver/Pdo/AbstractPdoConnection.php
@@ -101,13 +101,13 @@ abstract class AbstractPdoConnection extends AbstractConnection implements PdoCo
         $connectionContext = new ConnectionContext(__METHOD__);
 
         try {
-            $this->logger?->log(LogLevel::INFO, $token);
+            $this->logger?->log(LogLevel::INFO, $token, ['type' => LogType::CONNECTION]);
             $this->profiler?->begin($token, $connectionContext);
             $this->initConnection();
             $this->profiler?->end($token, $connectionContext);
         } catch (PDOException $e) {
             $this->profiler?->end($token, $connectionContext->setException($e));
-            $this->logger?->log(LogLevel::ERROR, $token);
+            $this->logger?->log(LogLevel::ERROR, $token, ['type' => LogType::CONNECTION]);
 
             throw new Exception($e->getMessage(), (array) $e->errorInfo, $e);
         }
@@ -119,6 +119,7 @@ abstract class AbstractPdoConnection extends AbstractConnection implements PdoCo
             $this->logger?->log(
                 LogLevel::DEBUG,
                 'Closing DB connection: ' . $this->driver->getDsn() . ' ' . __METHOD__,
+                ['type' => LogType::CONNECTION],
             );
 
             $this->pdo = null;
@@ -225,7 +226,7 @@ abstract class AbstractPdoConnection extends AbstractConnection implements PdoCo
             try {
                 $transaction->rollBack();
             } catch (Throwable $e) {
-                $this->logger?->log(LogLevel::ERROR, (string) $e, [__METHOD__]);
+                $this->logger?->log(LogLevel::ERROR, (string) $e, [__METHOD__, 'type' => LogType::TRANSACTION]);
                 /** hide this exception to be able to continue throwing original exception outside */
             }
         }