+++ tests/Db/Helper/DbStringHelperTest.php
@@ -18,6 +18,14 @@ final class DbStringHelperTest extends TestCase
         $this->assertSame('TestCase', DbStringHelper::baseName('TestCase'));
     }
 
+    public function testIsReadQuery(): void
+    {
+        $this->assertTrue(DbStringHelper::isReadQuery('SELECT * FROM tbl'));
+        $this->assertTrue(DbStringHelper::isReadQuery('SELECT * FROM tbl WHERE id=1'));
+        $this->assertTrue(DbStringHelper::isReadQuery('SELECT * FROM tbl WHERE id=1 LIMIT 1'));
+        $this->assertTrue(DbStringHelper::isReadQuery('SELECT * FROM tbl WHERE id=1 LIMIT 1 OFFSET 1'));
+    }
+
     public static function pascalCaseToIdProvider(): array
     {
         return [