+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Enh #801: Deprecate `SchemaInterface::isReadQuery()` and add `DbStringHelper::isReadQuery()` method (@Tigrov)
 - Enh #801: Remove unnecessary symbol `\\` from `rtrim()` function inside `DbStringHelper::baseName()` method (@Tigrov)
 - Bug #801: Fix bug with `Quoter::$tablePrefix` when change `AbstractConnection::$tablePrefix` property (@Tigrov)
+- Enh #809: Add psalm type for parameters to bind to the SQL statement (@vjik)
 - Enh #810: Add more specific psalm type for `QueryFunctionsInterface::count()` result (@vjik)
 
 ## 1.2.0 November 12, 2023